requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pandas
openpyxl
pyarrow
python-docx
python-dotenv
openai
